🏷️ Move hierarchy types to `types.hierarchy`
